Disable cnautilus executable and related functionality
The build for the cnautilus executable, including all of its dependencies, has been commented out in the CMakeLists.txt. This change also affects the properties and install directives for cnautilus. Furthermore, the testing commands in the pyproject.toml file and some options in the GitHub workflows file have been deactivated or changed.