Implement auto prediction of phosphate map
Added a functionality to automatically predict a phosphate map before building the model in 'nucleofind' build script. When the 'predin' argument is set to "auto", the script will run the 'predict_map' function and output a phosphate map into the current working directory. This feature enhances the capability of the 'nucleofind' tool while making it more user-friendly. Additionally, the relevant documentation is updated to reflect this change.