Add option to disable progress bar in Prediction class
An option has been added to the Prediction class, enabling users to disable the progress bar if they choose. This is particularly useful for instances when the progress bar could clutter the console output. The disabling of the progress bar can now be controlled by input argument in the command line.