Refactor GitHub workflows and adjust test parameters
The commit changes the structure of the GitHub workflows for increased transparency and ease of understanding. Tests are now run without the '-s' option in test.yml. Moreover, the actions in wheels.yml have been segmented and named for better clarity. The number of residues built in a unit test from 'test_building.py' has also been adjusted to better align with operational requirements.